[21.05] silence warnings when swap isn't there anymore

+++ nixos/infrastructure/flyingcircus-physical.nix
@@ -49,7 +49,9 @@ mkIf (cfg.infrastructureModule == "flyingcircus-physical") {
     flyingcircus.activationScripts = {
       disableSwap = ''
         swapoff -a
-        wipefs -af /dev/disk/by-label/swap || true
+        if [[ -e /dev/disk/by-label/swap ]]; then
+          wipefs  -af /dev/disk/by-label/swap || true
+        fi
       '';
     };
     systemd.targets.swap.enable = false;  # implicitly mask the unit to prevent pulling in existing `*.swap` units
